+++ static/js/shim.js
@@ -9,10 +9,6 @@ that still refer to the old name.
 var narrow_state = {}; // global, should be made into module
 narrow_state.set_compose_defaults = narrow.set_compose_defaults;
 
-var compose_actions = {};
-compose_actions.start = compose.start;
-compose_actions.cancel = compose.cancel;
-
 var compose_state = {};
 compose_state.has_message_content = compose.has_message_content;
 compose_state.recipient = compose.recipient;
